move: api 폴더 내부의 js 파일 경로 수정.
